Reorganize models in core.models
