+++ server/core/views/api.py
@@ -80,10 +80,10 @@ class ClassCatalog(APIView):
 def get_student_dashboard(request):
     user = request.user
 
-    previous_programs = Program.get_previous_student_programs(user)
+    previous_programs = StudentRegistration.get_previous_programs(user)
     previous_json = [{"name": str(p), "url": p.url} for p in previous_programs]
 
-    current_programs = Program.get_current_student_programs(user)
+    current_programs = StudentRegistration.get_current_programs(user)
 
     open_programs = Program.get_open_student_programs().exclude(id__in=current_programs)
     # TODO add grade checks
